Nonce should never be 0 bytes
At OpenRelay, we're trying to use Ganache to test our services that
are written in Go. The RPC client that ships with Geth ignores the
hash provided by RPC and recalculates it instead. Due to some discrepancies
between the value used to calculate the hash and the values served over
RPC, the Go client is unable to compute the same hash as Ganache.

One of the issues we ran into was that ethereumjs-block defaults to
an empty nonce, while according to the spec a nonce should be 8 bytes.
When computing the hash, the difference between a 0 byte nonce and
a nonce with 8 null bytes is obviously significant.

With this change along with a couple of changes to Ganache's RPC
encoding, the Ethereum client within Geth computes the same block
hashes as Ganache.