docs: mention that stack in step event can change
The step event provides data obj which includes `stack` and it is a reference of vm's internal stack obj. In next steps, this stack object changes and if external code had snapshotted the stack's reference instead of deep cloning, the snapshot may point to the latest version of stack instead of one that was snapshotted.