fixed kademlia kbucket add
in the kbucket implementation the dontSplit attribute of KBucketNode is typed as bool, but was checked wit "!== undefined" which always was true. This caused KBucket.add to never split the leaf node and therefore kept the kbucket as a single level tree with a maximum of _numberOfNodesPerKBucket entrys in total.